Tests are run correctly at compile time when using gsc
Created a Makefile to demonstrate how it is used.

Changed if expressions to be like smalltalk-80.  I plan
to modify gambit or make a preprocessor which which have
syntactic sugar for lambdas.  As I want to use the lambda syntax
anywhere deferred execution is used, I want my own if/cond
expressions