In an effort to make bug more in the style of Donald Knuth's "literate programming", I moved all of the code into one file. Since code can be executed at compile time without affecting the resulting executable in Gambit C, any external files which I must produce (such as files for macros, or namespace files) can be generated from the main program itself at compile time.
I think there are some macros waiting to be discovered to make
the syntax for this simpler, but it is functional for now at
least.

Also, decided that I should not define names which conflict
with Gambit's names if I ever intend to use those procedures
from Gambit.  For instance, I prefer my "if" always,  so
I have no problem setting the default namespace for "if"
to be lang.  However, I plan to use cons/car/cdr, so
don't override those names for streams.

Also, updated the build system to reflect this change, and
cleaned up "make clean"