Changed the naming standards so that regular text editors will highlight BUG files as scheme files by default.
Updated Makefiles so that they stop complaining about datarootdir.

Remove repetition in "include"s which write out to file

Macros are working in a REPL, regardless of namespace of
the calling code.  Working for my cases, but the code is ugly
and doesn't cover all cases.
For instance, when expanding out the macros to be written to file,
I've assumed that the form is `(foo bar baz), and as such I remove
the quasiquote.  This needs to be more generic.